refactor: address code review feedback and update documentation
- Change host flag from -h to -H to avoid conflict with help
- Move is_local_path and parse_s3_path to utils module for better organization
- Add configurable backup interval via BGPKIT_BROKER_BACKUP_INTERVAL_HOURS
- Improve error message for temporary directory creation
- Update configuration summary to show backup interval
- Update README.md with comprehensive backup and environment variable documentation
- Update CHANGELOG.md to document configurable backup interval
